Postprocess node visitor refactor
